CBIS-DDSM/Code/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
Pillow>=9.0.0
matplotlib>=3.4.0
seaborn>=0.13.2
